Cups and Targets
Added 3D models for basic cups and connected to image targets. Image
targets will now no longer be white planes and will show a target image.
This build has little to no issues and is stable.

Added APK. If a person with an andriod phone wanted to use the app, they
would have to download this APK onto their phone and run it.